requirements.txt
Django==6.0.4
gunicorn==23.0.0
whitenoise==6.9.0
Pillow==12.2.0
dj-database-url==2.3.0
sqlparse==0.5.5
asgiref==3.11.1
tzdata==2026.1
